+++ apps/indexer-proxy/proxy/src/p2p.rs
@@ -72,7 +72,7 @@ pub async fn stop_network() {
     }
 }
 
-pub async fn report_conflict(deployment: &str, channel: &str, conflict: i32, start: i64, end: i64) {
+pub async fn report_conflict(deployment: &str, channel: &str, conflict: u64, start: i64, end: i64) {
     let senders = P2P_SENDER.read().await;
     if senders.is_empty() {
         warn!("NONE NETWORK WHEN REPORT CONFLICT");
